proxy/requirements.txt
mitmproxy
python-magic
pymongo
websockets
asyncio
grpcio
grpcio-tools